A wrapper of future_lapply that has the correct environment

Fixes future_lapply issue where the environment is not set properly.

futurenow_lapply(
  X,
  FUN,
  ...,
  future.envir = parent.frame(),
  future.globals = TRUE,
  future.packages = NULL,
  future.lazy = FALSE,
  future.seed = FALSE,
  future.scheduling = 1,
  future.chunk.size = NULL,
  future.label = "futurenow_lapply-%d"
)

Arguments

X
Vector or list, R objects

FUN
Function to apply on each element of X

...
Further argument to be passed to FUN or future_eapply

future.envir
Where the future instances should be created

future.globals
A logical, a character vector, or a named list for controlling how global variables are handled; See future_eapply

future.packages
Character vector of packages for future instances to load

future.lazy
Whether to start evaluation immediately

future.seed
A logical or an integer (of length one or seven), or a list of length(X) with pre-generated random seeds; see future_eapply

future.scheduling, future.chunk.size
See future_eapply

future.label
Label to be assigned to each future instances

Value

A list with same length and names as X. See lapply for details.

See also